# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ConectApps Rest API",
        "description": "ConectApps Rest API",
        "contact": {
            "email": "loisagaston@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a credential User."
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name Field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create User credentials"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "responses": {
                    "200": {
                        "description": "Mostrar credencial User."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh",
                "description": "Refresh",
                "operationId": "7d5da2870d3eba95cdd67878ab20da3a",
                "responses": {
                    "200": {
                        "description": "Mostrar credencial User."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/posts/pdf": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Muestra toda la información almacenada en base de datos y la retorna en forma de tabla en un archivo PDF.",
                "description": "Display a listing of the resource.\nMostramos el listado de los regitros solicitados.",
                "operationId": "e369411d95fabf37e0260988cd2bcf3f",
                "responses": {
                    "200": {
                        "description": "Muestra toda la información almacenada en base de datos y la retorna en forma de tabla en un archivo PDF."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/post": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Crea un nuevo post",
                "description": "Retorna data de post",
                "operationId": "storePost",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Muestra información almacenada en base de datos filtrando por id y en formato json.",
                "description": "Display the specified resource.\nMuestra el registro solicitado",
                "operationId": "e051e88af1a8c62e43ce3f1495236469",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta de datos de un post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id de post.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar info de un post."
                    },
                    "404": {
                        "description": "No se ha encontrado el post."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Elimina post",
                "description": "Elimina post",
                "operationId": "5425ebaf55acd55d9af4945aa2d5b57d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}